Refactor to move snekbox to module
