+++ bot/exts/utils/snekbox/_cog.py
@@ -3,10 +3,8 @@ from __future__ import annotations
 import asyncio
 import contextlib
 import re
-from dataclasses import dataclass, field
 from functools import partial
 from operator import attrgetter
-from signal import Signals
 from textwrap import dedent
 from typing import Literal, Optional, TYPE_CHECKING, Tuple
 
@@ -19,7 +17,7 @@ from bot.bot import Bot
 from bot.constants import Channels, MODERATION_ROLES, Roles, URLs
 from bot.decorators import redirect_output
 from bot.exts.help_channels._channel import is_help_forum_post
-from bot.exts.utils.snekio import FILE_SIZE_LIMIT, FileAttachment, sizeof_fmt
+from bot.exts.utils.snekbox._eval import EvalJob, EvalResult
 from bot.log import get_logger
 from bot.utils import send_to_paste_service
 from bot.utils.lock import LockedResourceError, lock_arg
@@ -81,113 +79,10 @@ NO_SNEKBOX_CHANNELS = (Channels.python_general,)
 NO_SNEKBOX_CATEGORIES = ()
 SNEKBOX_ROLES = (Roles.helpers, Roles.moderators, Roles.admins, Roles.owners, Roles.python_community, Roles.partners)
 
-SIGKILL = 9
-
 REDO_EMOJI = '\U0001f501'  # :repeat:
 REDO_TIMEOUT = 30
 
 PythonVersion = Literal["3.10", "3.11"]
-
-
-@dataclass
-class EvalJob:
-    """Job to be evaluated by snekbox."""
-
-    args: list[str]
-    files: list[FileAttachment] = field(default_factory=list)
-    name: str = "eval"
-    version: PythonVersion = "3.11"
-
-    @classmethod
-    def from_code(cls, code: str, path: str = "main.py") -> EvalJob:
-        """Create an EvalJob from a code string."""
-        return cls(
-            args=[path],
-            files=[FileAttachment(path, code.encode())],
-        )
-
-    def as_version(self, version: PythonVersion) -> EvalJob:
-        """Return a copy of the job with a different Python version."""
-        return EvalJob(
-            args=self.args,
-            files=self.files,
-            name=self.name,
-            version=version,
-        )
-
-    def to_dict(self) -> dict[str, list[str | dict[str, str]]]:
-        """Convert the job to a dict."""
-        return {
-            "args": self.args,
-            "files": [file.to_dict() for file in self.files],
-        }
-
-
-@dataclass(frozen=True)
-class EvalResult:
-    """The result of an eval job."""
-
-    stdout: str
-    returncode: int | None
-    files: list[FileAttachment] = field(default_factory=list)
-    err_files: list[str] = field(default_factory=list)
-
-    @property
-    def status_emoji(self) -> str:
-        """Return an emoji corresponding to the status code or lack of output in result."""
-        # If there are attachments, skip empty output warning
-        if not self.stdout.strip() and not self.files:  # No output
-            return ":warning:"
-        elif self.returncode == 0:  # No error
-            return ":white_check_mark:"
-        else:  # Exception
-            return ":x:"
-
-    def get_message(self, job: EvalJob) -> tuple[str, str]:
-        """Return a user-friendly message and error corresponding to the process's return code."""
-        msg = f"Your {job.version} {job.name} job has completed with return code {self.returncode}"
-        error = ""
-
-        if self.returncode is None:
-            msg = f"Your {job.version} {job.name} job has failed"
-            error = self.stdout.strip()
-        elif self.returncode == 128 + SIGKILL:
-            msg = f"Your {job.version} {job.name} job timed out or ran out of memory"
-        elif self.returncode == 255:
-            msg = f"Your {job.version} {job.name} job has failed"
-            error = "A fatal NsJail error occurred"
-        else:
-            # Try to append signal's name if one exists
-            with contextlib.suppress(ValueError):
-                name = Signals(self.returncode - 128).name
-                msg = f"{msg} ({name})"
-
-        # Add error message for failed attachments
-        if self.err_files:
-            failed_files = f"({', '.join(self.err_files)})"
-            msg += (
-                f".\n\n> Some attached files were not able to be uploaded {failed_files}."
-                f" Check that the file size is less than {sizeof_fmt(FILE_SIZE_LIMIT)}"
-            )
-
-        return msg, error
-
-    @classmethod
-    def from_dict(cls, data: dict[str, str | int | list[dict[str, str]]]) -> EvalResult:
-        """Create an EvalResult from a dict."""
-        res = cls(
-            stdout=data["stdout"],
-            returncode=data["returncode"],
-        )
-
-        for file in data.get("files", []):
-            try:
-                res.files.append(FileAttachment.from_dict(file))
-            except ValueError as e:
-                log.info(f"Failed to parse file from snekbox response: {e}")
-                res.err_files.append(file["path"])
-
-        return res
 
 
 class CodeblockConverter(Converter):
@@ -622,8 +517,3 @@ def predicate_message_edit(ctx: Context, old_msg: Message, new_msg: Message) -> 
 def predicate_emoji_reaction(ctx: Context, reaction: Reaction, user: User) -> bool:
     """Return True if the reaction REDO_EMOJI was added by the context message author on this message."""
     return reaction.message.id == ctx.message.id and user.id == ctx.author.id and str(reaction) == REDO_EMOJI
-
-
-async def setup(bot: Bot) -> None:
-    """Load the Snekbox cog."""
-    await bot.add_cog(Snekbox(bot))
