+++ bot/exts/utils/snekbox/_io.py
